Update Connect4.java
I made addChip return an integer instead. It should still work the same. return false is made into return -1.  If it does not return -1 and instead successfully adds the chip, the method returns the row the chip was added to. This is so the row and column could be gotten in Connect4Test so that I can call the isFinished method correctly.